hot fix

+++ www/console/controllers/MiCacheController.php
@@ -10,10 +10,14 @@ class MiCacheController extends \yii\console\Controller
     public function actionClear($key, $domain='miduoduo')
     {
         $cache = Yii::$app->global_cache;
-        $cache->keyPrefix = $domain;
-        $cache->delete($key);
-        echo "clear Done ! \n";
-        return true;
+        $cache->keyPrefix = $domain . '@';
+        if ($cache->exists($key)){
+            $cache->delete($key);
+            echo "clear Done ! \n";
+            return true;
+        } else {
+            echo "No cache found ! \n";
+        }
     }
 
 }
